requirements.txt
openai==1.6.1
chromadb==0.4.22
python-dotenv==1.0.0
streamlit==1.31.1